requirements.txt
numpy==1.16.2
Pillow==5.4.1
pygame==1.9.4
